requirements.txt
streamlit>=1.30.0
reportlab>=4.0
googlemaps>=4.10.0
anthropic>=0.40.0
requests>=2.31.0
python-dotenv>=1.0.0
